refactoring: dto and entity packaged combined to model package in security
