+++ main/src/main/kotlin/com/ip13/main/model/entity/BookingConstraint.kt
@@ -1,6 +1,6 @@
 package com.ip13.main.model.entity
 
-import com.ip13.main.security.entity.User
+import com.ip13.main.security.model.entity.User
 import jakarta.persistence.Entity
 import jakarta.persistence.GeneratedValue
 import jakarta.persistence.GenerationType
