+++ main/src/main/kotlin/com/ip13/main/model/entity/Restaurant.kt
@@ -1,7 +1,7 @@
 package com.ip13.main.model.entity
 
 import com.fasterxml.jackson.annotation.JsonIgnore
-import com.ip13.main.security.entity.User
+import com.ip13.main.security.model.entity.User
 import jakarta.persistence.*
 
 @Entity
