+++ main/src/main/kotlin/com/ip13/main/security/model/dto/RequestDto.kt
@@ -1,4 +1,4 @@
-package com.ip13.main.security.dto
+package com.ip13.main.security.model.dto
 
 import com.ip13.main.model.validation.Password
 import jakarta.validation.constraints.NotBlank
