+++ main/src/main/kotlin/com/ip13/main/security/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.ip13.main.security.repository
 
-import com.ip13.main.security.entity.User
+import com.ip13.main.security.model.entity.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
