+++ main/src/test/kotlin/com/ip13/main/service/BookingConstraintServiceTest.kt
@@ -5,7 +5,7 @@ import com.ip13.main.model.dto.request.AddBookingConstraintRequestDto
 import com.ip13.main.model.entity.BookingConstraint
 import com.ip13.main.model.entity.Restaurant
 import com.ip13.main.repository.BookingConstraintRepository
-import com.ip13.main.security.entity.User
+import com.ip13.main.security.model.entity.User
 import com.ip13.main.security.service.UserService
 import io.mockk.every
 import io.mockk.impl.annotations.InjectMockKs
